ot-change-validation-tool/requirements.txt
# OT Change Validation Tool - Requirements

# Web UI
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0

# HTTP requests
requests>=2.31.0

# API (optional REST API)
fastapi>=0.104.0
uvicorn>=0.24.0

# Scheduling
apscheduler>=3.10.0

# Email parsing (for ID alerts)
imapclient>=2.3.0

# Environment management
python-dotenv>=1.0.0

# Excel export for reports
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# CLI
click>=8.1.0
rich>=13.0.0
